requirements.txt
streamlit
pandas
gspread
oauth2client
pillow
pyzbar
opencv-python-headless